+++ packages/hardhat-graph-protocol/CHANGELOG.md
@@ -1,5 +1,19 @@
 # hardhat-graph-protocol
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphprotocol/toolshed@0.4.2
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphprotocol/toolshed@0.4.1
+
 ## 0.1.13
 
 ### Patch Changes